backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
fastapi-cors
sqlalchemy
openai-whisper
ffmpeg-python
python-multipart
ollama
chromadb
sentence-transformers
yt-dlp
openai
anthropic
google-generativeai
gTTS
appdirs
pycryptodome